+++ CloudShell.Abstractions/ResourceManager/ResourcePredefinedViewIds.cs
@@ -11,6 +11,7 @@ public static class ResourcePredefinedViewIds
     public static ResourceViewId Volumes { get; } = new(ResourceTabGroupIds.Storage, "volumes");
     public static ResourceViewId Activity { get; } = new(ResourceTabGroupIds.Management, "activity");
     public static ResourceViewId Health { get; } = new(ResourceTabGroupIds.Management, "health");
+    public static ResourceViewId Recovery { get; } = new(ResourceTabGroupIds.Management, "recovery");
     public static ResourceViewId Monitoring { get; } = new(ResourceTabGroupIds.Management, "monitoring");
     public static ResourceViewId Environment { get; } = new(ResourceTabGroupIds.Management, "environment");
     public static ResourceViewId Logs { get; } = new(ResourceTabGroupIds.Telemetry, "logs");